% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_manual_ipa.R
\name{add_manual_ipa}
\alias{add_manual_ipa}
\title{Manually Add a Word to the Phonetic Dictionary}
\usage{
add_manual_ipa(
  word,
  ipa,
  dict_path = "phonetic_dictionary.csv",
  overwrite = TRUE
)
}
\arguments{
\item{word}{a single word}

\item{ipa}{the IPA string representing the phonetic spelling of that word}

\item{dict_path}{path to the local phonetic dictionary csv (default = "phonetic_dictionary.csv")}
}
\value{
an updated tibble with the new entry
}
\description{
Adds a custom word and IPA transcription to your local phonetic dictionary.
}
