% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_dictionary.R
\name{add_to_dictionary}
\alias{add_to_dictionary}
\title{Manually Add Words to Phonetic Dictionary}
\usage{
add_to_dictionary(words, dict_path = "phonetic_dictionary.csv")
}
\arguments{
\item{words}{a character vector of one or more words to add}

\item{dict_path}{path to the local phonetic dictionary (defaults to "phonetic_dictionary.csv")}
}
\value{
an updated tibble with all dictionary entries
}
\description{
Allows users to manually add one or more words to their local phonetic dictionary
}
