% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_ipa.R
\name{clean_ipa}
\alias{clean_ipa}
\title{Convert Text to IPA}
\usage{
clean_ipa(ipa)
}
\arguments{
\item{sentence}{a regular text string that you want to convert to phonetic spellings}

\item{dict_path}{path to the local phonetic dictionary file}
}
\value{
a tibble with words from the text string and their phonetic spellings; will return "NA" for words not found in dictionary.
}
\description{
Converts text strings to phonetic spelling using the international phonetic alphabet (IPA)
}
