% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_phonetic_alignment.R
\name{compare_phonetic_alignment}
\alias{compare_phonetic_alignment}
\title{Compare Phonetic Alignment}
\usage{
compare_phonetic_alignment(text1, text2, dict_path = "phonetic_dictionary.csv")
}
\arguments{
\item{text2}{a text string for the second sentence}

\item{dict_path}{path to the local phonetic dictionary file}

\item{text}{1 a text string for the first sentence}
}
\value{
a list with:
\describe{
\item{total_errors}{the number of total errors}
\item{substitutions}{the number of phonetic substitutions}
\item{deletions}{the number of phonetic character deletions}
\item{insertions}{the numnber of phonetic character insertions}
\item{percent_misaligned}{the percent of the sequence that doesn't match}
\item{aligned_seq1}{the aligned IPA sequence for the first text string}
\itme{aligned_seq2}{the aligned IPA sequence for the second text string}
}
}
\description{
Function to generate IPA versions of two text strings for comparison with tally of errors, error types (substitution, deletion, insertion), and misalignment percentage.
}
