% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dictionary_from_csv.R
\name{generate_dictionary_from_csv}
\alias{generate_dictionary_from_csv}
\title{Generate Phonetic Dictionary from CSV}
\usage{
generate_dictionary_from_csv(
  csv_path,
  text_columns,
  out_path = "phonetic_dictionary.csv"
)
}
\arguments{
\item{csv_path}{path to user's .csv file.}

\item{text_columns}{character vector of columns containing text to be imported}

\item{out_path}{path to create or update local phonetic dictionary (defaults to "phonetic_dictionary.csv")}
}
\value{
a tibble that contains unique words and their IPA spellings.
}
\description{
Function to extract all unique words from one or more columns in a .csv document.
}
