% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_word_in_dictionary.R
\name{replace_word_in_dictionary}
\alias{replace_word_in_dictionary}
\title{Replace the IPA Spelling for a Word in the Dictionary}
\usage{
replace_word_in_dictionary(
  word,
  old_spelling,
  new_spelling,
  dict_path = "phonetic_dictionary.csv"
)
}
\arguments{
\item{word}{the word for which the spelling is being replaced}

\item{old_spelling}{the current ipa spelling that needs to be replaced}

\item{new_spelling}{the manually updated ipa spelling}

\item{dict_path}{path to the local phonetic dictionary}
}
\value{
the word and its updated spelling
}
\description{
Updates the IPA spelling of an existing word in the local phonetic dictionary.
}
