% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_phonetic_dictionary.R
\name{scrape_bulk_phonetic_spelling}
\alias{scrape_bulk_phonetic_spelling}
\title{Scrape Dictionary}
\usage{
scrape_bulk_phonetic_spelling(word_list, out_path = "phonetic_dictionary.csv")
}
\arguments{
\item{word_list}{a character vector of words to scrape}

\item{out_path}{path to local phonetic dictionary csv file; will default to "phonetic_dictionary.csv"}
}
\value{
a .csv with words and their phonetic spellings
}
\description{
Scrapes Wiktionary for the International Phonetic Alphabet (IPA) spelling of words
}
