% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_phonemes.R
\name{tokenize_phonemes}
\alias{tokenize_phonemes}
\title{Tokenize IPA Strings into Phonemes}
\usage{
tokenize_phonemes(ipa_string)
}
\arguments{
\item{ipa_string}{a cleaned IPA string}
}
\value{
character vector of phonemes
}
\description{
Internal function that breaks strings into phonemes using regex rules. (English)
}
